# Insert data to test performance and tunning queries