--- a/queries.sql
+++ b/queries.sql
@@ -209,3 +209,8 @@ WHERE V.name = 'Maisy Smith'
 GROUP BY species
 ORDER BY animal_frec DESC
 LIMIT 1;
+
+/* EXPLAIN ANALYZE QUERIES */
+explain analyze SELECT COUNT(*) FROM visits where animals_id = 4;
+explain analyze SELECT * FROM visits where vets_id = 2;
+explain analyze SELECT * FROM owners where email = 'owner_18327@mail.com';